_legacy/requirements-complete.txt
# =============================================================================
# COMPLETE PROJECT REQUIREMENTS
# Blood Test Analysis System - Python Backend + Next.js Frontend
# =============================================================================

# =============================================================================
# PYTHON BACKEND DEPENDENCIES
# =============================================================================

# Core Python packages
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Google API packages
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1

# PDF processing
PyMuPDF==1.26.4
pdfminer.six==20250506
pdfplumber==0.11.7
PyPDF2==3.0.1
pypdfium2==4.30.0

# Image processing
Pillow==11.3.0

# Data processing
pandas==2.3.2
numpy==2.3.3

# OpenAI API
openai==1.107.3

# HTTP and networking
requests==2.32.5
requests-oauthlib==2.0.0
httplib2==0.31.0
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0

# Authentication and security
cryptography==46.0.0
oauthlib==3.3.1
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Data validation and parsing
pydantic==2.11.9
pydantic_core==2.33.2
protobuf==6.32.1
proto-plus==1.26.1

# Utilities
tqdm==4.67.1
cachetools==5.5.2
distro==1.9.0
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# Text processing
charset-normalizer==3.4.3
idna==3.10
six==1.17.0

# Async support
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
jiter==0.11.0

# URL handling
uritemplate==4.2.0

# SSL and certificates
certifi==2025.8.3

# C extensions
cffi==2.0.0
pycparser==2.23

# Parsing
pyparsing==3.2.4

# =============================================================================
# NODE.JS FRONTEND DEPENDENCIES
# =============================================================================
# Note: These are listed for reference. Install with: npm install
# See web/package.json for exact versions

# Core Framework
# next@15.5.3
# react@19.1.0
# react-dom@19.1.0
# typescript@5

# UI and Styling
# tailwindcss@4
# @tailwindcss/postcss@4
# tw-animate-css@1.3.8
# tailwind-merge@3.3.1
# class-variance-authority@0.7.1
# clsx@2.1.1

# UI Components
# @radix-ui/react-dialog@1.1.15
# @radix-ui/react-dropdown-menu@2.1.16
# @radix-ui/react-label@2.1.7
# @radix-ui/react-select@2.2.6
# @radix-ui/react-separator@1.1.7
# @radix-ui/react-slot@1.2.3
# @radix-ui/react-switch@1.2.6
# @radix-ui/react-tooltip@1.2.8

# Charts and Visualization
# recharts@3.2.1

# Drag and Drop
# @dnd-kit/core@6.3.1
# @dnd-kit/sortable@10.0.0
# @dnd-kit/utilities@3.2.2

# Icons and Utilities
# lucide-react@0.544.0
# next-themes@0.4.6

# Google APIs
# googleapis@160.0.0

# Development Dependencies
# @eslint/eslintrc@3
# eslint@9
# eslint-config-next@15.5.3
# @types/node@20
# @types/react@19
# @types/react-dom@19

# Testing
# jest@30.1.3
# jest-environment-jsdom@30.1.2
# @testing-library/jest-dom@6.8.0
# @testing-library/react@16.3.0
# @testing-library/user-event@14.6.1

# =============================================================================
# INSTALLATION INSTRUCTIONS
# =============================================================================

# Python Backend:
# pip install -r requirements.txt

# Node.js Frontend:
# cd web
# npm install

# =============================================================================
# ENVIRONMENT VARIABLES REQUIRED
# =============================================================================

# Python Backend (config.py or .env):
# OPENAI_API_KEY=your_openai_key
# GOOGLE_CREDENTIALS_FILE=path_to_credentials.json
# DRIVE_FOLDER_ID=your_drive_folder_id
# SHEET_ID=your_google_sheet_id

# Web Frontend (.env.local):
# GOOGLE_SHEETS_SPREADSHEET_ID=your_sheet_id
# GOOGLE_SERVICE_ACCOUNT_EMAIL=your_service_account_email
# GOOGLE_SERVICE_ACCOUNT_KEY=your_private_key
# NEXT_PUBLIC_LOGIN_USERNAME=your_username
# NEXT_PUBLIC_LOGIN_PASSWORD=your_password
